add two new routes, refactor
