+++ app/controllers/api_controller.rb
@@ -3,8 +3,9 @@
 require 'benchmark'
 
 class ApiController < ApplicationController
-  before_action :set_arguments, only: [:call]
-  before_action :set_endpoint, only: [:call]
+  include ApiHelper
+  before_action :set_arguments, only: [:call, :validate_params, :validate_param]
+  before_action :set_endpoint, only: [:call, :validate_params, :validate_param]
   before_action :set_endpoints, only: [:list_endpoints]
 
   # returns list of endpoints
@@ -28,29 +29,46 @@ class ApiController < ApplicationController
     end
   end
 
+  # validates all params relative to endpoint expectations
+  def validate_params
+    if @endpoint.nil?
+      error_string = "no endpoint was found: #{@client_tag} with request #{@request_name}"
+      render json: { error: error_string }, status: 404
+    else
+      begin
+        erb = EndpointRequestBuilder.new(endpoint)
+        erb.validate(@arguments)
+        render json: { success: "Params look great!"}, status: 200
+      rescue StandardError => e
+        render json: { error: e.to_s }, status: 500
+      end
+    end
+  end
+
+  # validates one param in isolation
+  def validate_param
+    if @endpoint.nil?
+      error_string = "no endpoint was found: #{@client_tag} with request #{@request_name}"
+      render json: { error: error_string }, status: 404
+    elsif !@endpoint.params.keys.include?(@arguments.keys.first)
+       error_string = "supplied argument '#{@arguments.keys.first}' is not a parame for the provided endpoint"
+       render json: { error: error_string }, status: 404
+    else
+      begin
+        erb = EndpointRequestBuilder.new(endpoint)
+        erb.validate_param(@arguments)
+        render json: { success: "Param looks great!"}, status: 200
+      rescue StandardError => e
+        render json: { error: e.to_s }, status: 500
+      end
+    end
+  end
+
   private
 
-  # executes the api request
   def make_request
     endpoint_client = EndpointClient.new(@endpoint)
     endpoint_client.request(@arguments)
   end
 
-  # endpoint is defined by its client and name
-  def set_endpoint
-    @endpoint = Endpoint.where(name: params[:request_name])
-                        .where(client_tag: params[:client_tag]).first
-  end
-
-  # list of endpoints like params if params supplied
-  # nil params yield all endpoints
-  def set_endpoints
-    @endpoints = Endpoint.where('client_tag like ?', "%#{params[:client_tag]}%")
-                         .where('name like ?', "%#{params[:request_name]}%")
-  end
-
-  # array of arguments to be used in client call
-  def set_arguments
-    @arguments = params[:arguments]
-  end
 end
